refactor: move participant data filtering to match component
