+++ app/components/Matches/GameInfo.jsx
@@ -26,25 +26,12 @@ const GameInfo = (props) => {
   };
 
   const displayDate = getMatchDate();
-  const participants = props.details.participantIdentities;
-
-  // eslint-disable-next-line consistent-return
-  const participantId = (() => {
-    for (let i = 0; i < participants.length; i += 1) {
-      if (participants[i].player.summonerId === props.user.id) {
-        return participants[i].participantId;
-      }
-    }
-  })();
-
-  const participantData = props.details.participants[participantId - 1];
-  const gameOutcome = participantData.stats.win ? 'Victory' : 'Defeat';
 
   return (
     <div>
       <h4>{queueMode[props.queue].map}</h4>
       <img src={`/images/champions/${champ}.png`} alt={`${champ} icon`} />
-      <h4>{gameOutcome}</h4>
+      <h4>{props.gameOutcome}</h4>
       <h4>{displayDate}</h4>
     </div>
   );
@@ -54,6 +41,7 @@ GameInfo.propTypes = {
   champion: PropTypes.number.isRequired,
   timestamp: PropTypes.number.isRequired,
   queue: PropTypes.number.isRequired,
+  gameOutcome: PropTypes.string.isRequired,
 };
 
 const mapStateToProps = ({ user }) => ({ user });
